# docs/openapi.yaml
openapi: 3.2.1
info:
  title: OpenMediaDownloaderIOS API
  version: 0.5.0
servers:
  - url: http://localhost:8080
paths:
...
  /v1/jobs:
    post:
      operationId: createJob
      summary: Queue a download job
      description: >-
        Normalizes one public post URL and checks it against the allowed platforms before
        queuing. Profiles, channels, playlists, feeds, and collections are rejected. Each API
        key may have a limited number of queued or running jobs.
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
                  description: >-
                    One public post on an allowed platform; profiles, channels, playlists, feeds, and
                    collections are rejected. Tracking parameters are removed, and an Instagram post
                    link that starts with the account handle is stored without the handle.
              additionalProperties: false
            example:
              url: https://vimeo.com/123
      responses:
        "202":
          description: The job was queued.
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/TraceId"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            Location:
              description: Path of the created job.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnsupportedURL"
        "429":
          $ref: "#/components/responses/TooManyJobs"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  headers:
    NoStore:
      description: Responses must not be stored by clients or caches.
      schema:
        type: string
        const: no-store
    TraceId:
      description: >-
        Identifier of this request in the server logs. Include it when reporting a problem;
        it reveals nothing about the request.
      schema:
        type: string
        pattern: "^[a-z2-7]{26}$"
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: "API key created with `omdi keys create`, sent as `Authorization: Bearer omdi_<id>_<secret>`."
  responses:
    InvalidRequest:
      description: The body is not a single JSON object with a string `url` and no other fields.
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: invalid_request
    Unauthorized:
      description: The API key is missing, malformed, unknown, or revoked.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
        WWW-Authenticate:
          schema:
            type: string
            const: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unauthorized
...
    RequestTooLarge:
      description: The body exceeds the configured request size limit.
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: request_too_large
    UnsupportedMediaType:
      description: The request content type is not `application/json`.
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unsupported_media_type
    UnsupportedURL:
      description: The URL is not a public HTTPS URL on an allowed platform.
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unsupported_url
    TooManyJobs:
      description: The API key already has the maximum number of queued or running jobs.
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: too_many_jobs
    Internal:
      description: An internal error occurred. No details are disclosed.
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: internal
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - unauthorized
            - not_found
            - conflict
            - request_too_large
            - unsupported_media_type
            - unsupported_url
            - too_many_jobs
            - internal
      additionalProperties: false
    Job:
      type: object
      required:
        - id
        - status
        - url
        - platform
        - error
        - error_detail
        - created_at
        - updated_at
        - expires_at
      properties:
        id:
          type: string
        status:
          type: string
          enum:
            - queued
            - running
            - succeeded
            - failed
            - canceled
        url:
          type: string
          description: The normalized source URL.
        platform:
          type: string
          enum:
            - youtube
            - instagram
            - tiktok
            - x
            - reddit
            - vimeo
        error:
          type:
            - string
            - "null"
          enum:
            - unsupported_url
            - extraction_failed
            - too_large
            - timeout
            - internal
            - null
        error_detail:
          type:
            - string
            - "null"
          description: >-
            Why an `extraction_failed` job failed, derived from fixed patterns in the media tools'
            output; null for other errors. `blocked` is an anti-bot check, while
            `forbidden` is a refused request that may succeed later. `tool_error` means no known
            pattern matched, and `invalid_output` means the downloaded files did not pass
            validation. Jobs that fail with `forbidden`, `rate_limited`, or `network_error` have
            already been attempted twice.
          enum:
            - out_of_memory
            - login_required
            - blocked
            - forbidden
            - rate_limited
            - unavailable
            - geo_restricted
            - age_restricted
            - no_media
            - network_error
            - egress_denied
            - invalid_output
            - processing_failed
            - tool_error
            - null
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          description: When the job and its files are removed.
        bundle:
          description: Present for newly succeeded multi-item jobs; absent for single-item and legacy jobs.
          $ref: "#/components/schemas/Bundle"
        items:
          type: array
          description: Present only for succeeded jobs.
          items:
            $ref: "#/components/schemas/Item"
      additionalProperties: false
    Item:
      type: object
      required:
        - id
        - file_name
        - media_type
        - size_bytes
        - download_url
        - download_expires_at
      properties:
        id:
          type: string
        file_name:
          type: string
          description: Server-generated download file name.
        media_type:
          type: string
        size_bytes:
          type: integer
          minimum: 1
        download_url:
          type: string
          format: uri
          description: Absolute link built from OMDI_PUBLIC_URL; replaced on every poll.
        download_expires_at:
          type: string
          format: date-time
      additionalProperties: false
    Bundle:
      type: object
      required:
        - file_name
        - media_type
        - size_bytes
        - download_url
        - download_expires_at
      properties:
        file_name:
          type: string
          description: Server-generated ZIP download name.
        media_type:
          type: string
          const: application/zip
        size_bytes:
          type: integer
          minimum: 1
        download_url:
          type: string
          format: uri
          description: Opaque-token URL at /v1/bundles/{token}, built from OMDI_PUBLIC_URL and replaced on every poll. May be reused for HEAD and Range until replacement or expiry.
        download_expires_at:
          type: string
          format: date-time
      additionalProperties: false
...